Update spdlog to version 1.14.1

+++ xlive/Blam/Engine/Networking/NetworkMessageTypeCollection.cpp
@@ -214,7 +214,7 @@ void __stdcall handle_channel_message_hook(void* thisx, int network_channel_inde
 		if (peer_network_channel->is_channel_state_5())
 		{
 			s_team_change* received_data = (s_team_change*)packet;
-			LOG_TRACE_NETWORK(L"[H2MOD-CustomMessage] recieved on handle_channel_message_hook team_change: {}", received_data->team_index);
+			LOG_TRACE_NETWORK(L"[H2MOD-CustomMessage] recieved on handle_channel_message_hook team_change: {}", (int16)received_data->team_index);
 			user_interface_controller_set_desired_team_index(_controller_index_0, received_data->team_index);
 			user_interface_controller_update_network_properties(_controller_index_0);
 		}
